distance reverted to 8 bit in list

+++ src/client/gui/linked_list.h
@@ -5,7 +5,7 @@ typedef struct ListItem {
   struct ListItem* prev;
   struct ListItem* next;
   uint8_t angle;
-  uint16_t distance;
+  uint8_t distance;
 } ListItem;
 
 typedef struct ListHead {
@@ -21,7 +21,7 @@ void List_init(ListHead* head, int max_length);
 void List_set_max_length(ListHead* head, int max_length);
 
 //add item in the head of the list, if all ok return the added item else NULL
-ListItem* List_insert(ListHead* head, uint8_t angle, uint16_t distance);
+ListItem* List_insert(ListHead* head, uint8_t angle, uint8_t distance);
 
 //deletes all nodes of the list
 int List_reset(ListHead* head);
